+++ apps/desktop/src/components/onboarding/index.test.tsx
@@ -4,7 +4,7 @@ import { afterEach, describe, expect, it } from 'vitest'
 import { $desktopOnboarding, type DesktopOnboardingState, type OnboardingContext } from '@/store/onboarding'
 import type { OAuthProvider } from '@/types/hermes'
 
-import { Picker } from './desktop-onboarding-overlay'
+import { Picker } from '.'
 
 function provider(id: string, name = id): OAuthProvider {
   return {
